requirements.txt
streamlit
opencv-python
numpy
boto3
python-dotenv
Pillow
